+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## master
 
+- Added ability to require `pong` messages from clients (for better detection of broken connections). ([@palkan][])
+
+Use `--pong_timoout=<number of seconds to wait for ping>` option to enable this feature. **NOTE:** This option is only applied to clients using the extended Action Cable protocol.
+
 - Allow configuring ping interval and ping timestamp precision per connection. ([@palkan][])
 
 You can use URL query params to configure ping interval and timestamp precision per connection: